+++ core/app/[locale]/(default)/page.tsx
@@ -8,7 +8,6 @@ import { revalidate } from '~/client/revalidate-target';
 import { ProductCardCarousel } from '~/components/product-card-carousel';
 import { ProductCardCarouselFragment } from '~/components/product-card-carousel/fragment';
 import { Slideshow } from '~/components/slideshow';
-import { LocaleType } from '~/i18n/routing';
 
 const HomePageQuery = graphql(
   `
@@ -35,12 +34,12 @@ const HomePageQuery = graphql(
 );
 
 interface Props {
-  params: {
-    locale: LocaleType;
-  };
+  params: Promise<{ locale: string }>;
 }
 
-export default async function Home({ params: { locale } }: Props) {
+export default async function Home({ params }: Props) {
+  const { locale } = await params;
+
   setRequestLocale(locale);
 
   const t = await getTranslations('Home');